requirements.txt
edgartools>=2.0.0
gitpython>=3.1.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
